+++ google-cloud-bigtable/src/main/java/com/google/cloud/bigtable/data/v2/stub/BigtableClientContext.java
@@ -26,13 +26,16 @@ import com.google.auth.Credentials;
 import com.google.auth.oauth2.ServiceAccountJwtAccessCredentials;
 import com.google.cloud.bigtable.data.v2.BigtableDataSettings;
 import com.google.cloud.bigtable.data.v2.internal.JwtCredentialsWithAudience;
+import com.google.cloud.bigtable.data.v2.stub.metrics.BuiltinMetricsConstants;
 import com.google.cloud.bigtable.data.v2.stub.metrics.CustomOpenTelemetryMetricsProvider;
 import com.google.cloud.bigtable.data.v2.stub.metrics.DefaultMetricsProvider;
 import com.google.cloud.bigtable.data.v2.stub.metrics.ErrorCountPerConnectionMetricTracker;
 import com.google.cloud.bigtable.data.v2.stub.metrics.MetricsProvider;
 import com.google.cloud.bigtable.data.v2.stub.metrics.NoopMetricsProvider;
 import io.grpc.ManagedChannelBuilder;
+import io.grpc.opentelemetry.GrpcOpenTelemetry;
 import io.opentelemetry.api.OpenTelemetry;
+import io.opentelemetry.sdk.OpenTelemetrySdk;
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -50,6 +53,7 @@ public class BigtableClientContext {
   private static final Logger logger = Logger.getLogger(BigtableClientContext.class.getName());
 
   @Nullable private final OpenTelemetry openTelemetry;
+  @Nullable private final OpenTelemetrySdk internalOpenTelemetry;
   private final ClientContext clientContext;
 
   public static BigtableClientContext create(EnhancedBigtableStubSettings settings)
@@ -84,18 +88,32 @@ public class BigtableClientContext {
             ? ((InstantiatingGrpcChannelProvider) builder.getTransportChannelProvider()).toBuilder()
             : null;
 
-    ErrorCountPerConnectionMetricTracker errorCountPerConnectionMetricTracker = null;
+    @Nullable OpenTelemetrySdk internalOtel = null;
+    @Nullable ErrorCountPerConnectionMetricTracker errorCountPerConnectionMetricTracker = null;
+
+    // Internal metrics are scoped to the connections, so we need a mutable transportProvider,
+    // otherwise there is
+    // no reason to build the internal OtelProvider
+    if (transportProvider != null) {
+      internalOtel =
+          settings.getInternalMetricsProvider().createOtelProvider(settings, credentials);
+      if (internalOtel != null) {
+        // Set up per connection error count tracker if all dependencies are met:
+        // a configurable transport provider + otel
+        errorCountPerConnectionMetricTracker =
+            setupPerConnectionErrorTracer(builder, transportProvider, internalOtel);
+
+        // Configure grpc metrics
+        configureGrpcOtel(transportProvider, internalOtel);
+      }
+    }
 
     if (transportProvider != null) {
       // Set up cookie holder if routing cookie is enabled
       if (builder.getEnableRoutingCookie()) {
         setupCookieHolder(transportProvider);
       }
-      // Set up per connection error count tracker if OpenTelemetry is not null
-      if (openTelemetry != null) {
-        errorCountPerConnectionMetricTracker =
-            setupPerConnectionErrorTracer(builder, transportProvider, openTelemetry);
-      }
+
       // Inject channel priming if enabled
       if (builder.isRefreshingChannel()) {
         transportProvider.setChannelPrimer(
@@ -117,12 +135,43 @@ public class BigtableClientContext {
           clientContext.getExecutor());
     }
 
-    return new BigtableClientContext(clientContext, openTelemetry);
+    return new BigtableClientContext(clientContext, openTelemetry, internalOtel);
   }
 
-  private BigtableClientContext(ClientContext clientContext, OpenTelemetry openTelemetry) {
+  private static void configureGrpcOtel(
+      InstantiatingGrpcChannelProvider.Builder transportProvider, OpenTelemetrySdk otel) {
+
+    GrpcOpenTelemetry grpcOtel =
+        GrpcOpenTelemetry.newBuilder()
+            .sdk(otel)
+            .addOptionalLabel("grpc.lb.locality")
+            // Disable default grpc metrics
+            .disableAllMetrics()
+            // Enable specific grpc metrics
+            .enableMetrics(BuiltinMetricsConstants.GRPC_METRICS.keySet())
+            .build();
+
+    @SuppressWarnings("rawtypes")
+    ApiFunction<ManagedChannelBuilder, ManagedChannelBuilder> oldConfigurator =
+        transportProvider.getChannelConfigurator();
+
+    transportProvider.setChannelConfigurator(
+        b -> {
+          if (oldConfigurator != null) {
+            b = oldConfigurator.apply(b);
+          }
+          grpcOtel.configureChannelBuilder(b);
+          return b;
+        });
+  }
+
+  private BigtableClientContext(
+      ClientContext clientContext,
+      OpenTelemetry openTelemetry,
+      @Nullable OpenTelemetrySdk internalOtel) {
     this.clientContext = clientContext;
     this.openTelemetry = openTelemetry;
+    this.internalOpenTelemetry = internalOtel;
   }
 
   public OpenTelemetry getOpenTelemetry() {
@@ -136,6 +185,9 @@ public class BigtableClientContext {
   public void close() throws Exception {
     for (BackgroundResource resource : clientContext.getBackgroundResources()) {
       resource.close();
+    }
+    if (internalOpenTelemetry != null) {
+      internalOpenTelemetry.close();
     }
   }